pyproject.toml
[project]
name = "datanadhi-log"
version = "0.1.0"
description = "Data Nadhi Python logging SDK"
requires-python = ">=3.11"
authors = [
    { name = "Nishaanth S P", email = "nishaanth.sp@dataarena.space" }
]
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "uuid6>=1.0",
    "python-dotenv>=1.0.0",
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
    "orjson>=3.11.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.0.291"
]

[tool.ruff]
line-length = 88
target-version = "py311"
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "W", "N", "UP", "RET", "LOG007", "PLR2004"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
